refactor BaseRoomSession strategy
