+++ packages/core/src/types/callFabric.ts
@@ -1,3 +1,5 @@
+// TODO: Should be changed to "Fabric" rather than "CallFabric"
+
 import {
   CallState,
   CallConnect,
