+++ packages/js/src/BaseRoomSession.ts
@@ -1,107 +1,58 @@
 import {
-  connect,
-  Rooms,
-  extendComponent,
   BaseComponentContract,
-  BaseComponentOptions,
-  BaseConnectionContract,
-  VideoAuthorization,
-  LOCAL_EVENT_PREFIX,
-  validateEventsToSubscribe,
+  BaseConnectionState,
+  connect,
   EventEmitter,
-  SDKWorker,
-  VideoLayoutChangedEventParams,
 } from '@signalwire/core'
 import {
-  getDisplayMedia,
   BaseConnection,
   BaseConnectionOptions,
-  BaseConnectionStateEventTypes,
-  supportsMediaOutput,
   createSpeakerDeviceWatcher,
+  getDisplayMedia,
   getSpeakerById,
+  supportsMediaOutput,
+  MediaEventNames,
 } from '@signalwire/webrtc'
-import type {
-  RoomSessionObjectEvents,
-  CreateScreenShareObjectOptions,
-  AddDeviceOptions,
-  AddCameraOptions,
-  AddMicrophoneOptions,
-  BaseRoomInterface,
-  RoomMethods,
-  StartScreenShareOptions,
-  RoomSessionConnectionContract,
-  BaseRoomSessionJoinParams,
-  AudioElement,
-} from './utils/interfaces'
-import { SCREENSHARE_AUDIO_CONSTRAINTS } from './utils/constants'
-import { audioSetSpeakerAction } from './features/actions'
-import {
-  RoomSessionScreenShareAPI,
-  RoomSessionScreenShareConnection,
-  RoomSessionScreenShare,
-  RoomSessionScreenShareEvents,
-} from './RoomSessionScreenShare'
-import {
-  RoomSessionDeviceAPI,
-  RoomSessionDeviceConnection,
-  RoomSessionDevice,
-  RoomSessionDeviceEvents,
-} from './RoomSessionDevice'
-import * as workers from './video/workers'
 import {
   addOverlayPrefix,
   LocalVideoOverlay,
   OverlayMap,
 } from './VideoOverlays'
+import {
+  AudioElement,
+  BaseRoomSessionContract,
+  StartScreenShareOptions,
+} from './utils/interfaces'
+import { SCREENSHARE_AUDIO_CONSTRAINTS } from './utils/constants'
+import {
+  RoomSessionScreenShare,
+  RoomSessionScreenShareAPI,
+  RoomSessionScreenShareConnection,
+  RoomSessionScreenShareEvents,
+} from './RoomSessionScreenShare'
+import * as workers from './video/workers'
 
 export interface BaseRoomSession<
-  T,
-  TEvents extends EventEmitter.ValidEventTypes = RoomSessionObjectEvents
-> extends RoomMethods,
-    RoomSessionConnectionContract,
-    BaseComponentContract,
-    BaseConnectionContract<TEvents> {
-  /**
-   * Joins the room session.
-   */
-  join(options?: BaseRoomSessionJoinParams): Promise<T>
+  EventTypes extends EventEmitter.ValidEventTypes = BaseRoomSessionEvents
+> extends BaseRoomSessionContract,
+    BaseConnection<EventTypes>,
+    BaseComponentContract {}
 
-  /**
-   * Leaves the room. This detaches all the locally originating streams from the
-   * room.
-   */
-  leave(): Promise<void>
-}
+export interface BaseRoomSessionOptions extends BaseConnectionOptions {}
 
-export interface BaseRoomSessionOptions
-  extends BaseConnection<RoomSessionObjectEvents> {
-  eventsWatcher?: SDKWorker<RoomSessionConnection>
-}
-
-export class RoomSessionConnection<
-    TEvents extends EventEmitter.ValidEventTypes = RoomSessionObjectEvents
+export class BaseRoomSessionConnection<
+    EventTypes extends EventEmitter.ValidEventTypes = BaseRoomSessionEvents
   >
-  extends BaseConnection<TEvents>
-  implements BaseRoomInterface, RoomSessionConnectionContract
+  extends BaseConnection<EventTypes>
+  implements BaseRoomSessionContract
 {
   private _screenShareList = new Set<RoomSessionScreenShare>()
-  private _deviceList = new Set<RoomSessionDevice>()
   private _audioEl: AudioElement
   private _overlayMap: OverlayMap
   private _localVideoOverlay: LocalVideoOverlay
-  private _lastLayoutEvent: VideoLayoutChangedEventParams
 
-  constructor(options: BaseRoomSessionOptions) {
-    super(options)
-
-    this.initWorker()
-  }
-
-  protected initWorker() {
-    this.runWorker('videoWorker', {
-      worker: workers.videoWorker,
-    })
+  get audioEl() {
+    return this._audioEl
   }
 
   set overlayMap(map: OverlayMap) {
@@ -120,44 +71,72 @@ export class RoomSessionConnection<
     return this._localVideoOverlay
   }
 
-  set lastLayoutEvent(event: VideoLayoutChangedEventParams) {
-    this._lastLayoutEvent = event
-  }
-
-  get lastLayoutEvent() {
-    return this._lastLayoutEvent
-  }
-
-  get currentLayout() {
-    return this._lastLayoutEvent?.layout
-  }
-
-  get currentPosition() {
-    return this._lastLayoutEvent?.layout.layers.find(
-      (layer) => layer.member_id === this.memberId
-    )?.position
-  }
-
   get screenShareList() {
     return Array.from(this._screenShareList)
   }
 
-  get deviceList() {
-    return Array.from(this._deviceList)
-  }
+  private _attachSpeakerTrackListener() {
+    if (!supportsMediaOutput()) return
 
-  get interactivityMode() {
-    return this.select(({ session }) => {
-      const { authState } = session
-      return (authState as VideoAuthorization)?.join_as ?? ''
+    // @TODO: Stop the watcher when user leave/disconnects
+    createSpeakerDeviceWatcher().then((deviceWatcher) => {
+      deviceWatcher.on('removed', async (data) => {
+        const sinkId = this._audioEl.sinkId
+        const disconnectedSpeaker = data.changes.find((device) => {
+          const payloadDeviceId = device.payload.deviceId
+
+          return (
+            payloadDeviceId === sinkId ||
+            (payloadDeviceId === '' && sinkId === 'default') ||
+            (payloadDeviceId === 'default' && sinkId === '')
+          )
+        })
+        if (disconnectedSpeaker) {
+          this.emit('speaker.disconnected', {
+            deviceId: disconnectedSpeaker.payload.deviceId,
+            label: disconnectedSpeaker.payload.label,
+          })
+
+          /**
+           * In case the currently in-use speaker disconnects, OS by default fallbacks to the default speaker
+           * Set the sink id here to make the SDK consistent with the OS
+           */
+          await this._audioEl.setSinkId?.('')
+
+          const defaultSpeakers = await getSpeakerById('default')
+
+          if (!defaultSpeakers?.deviceId) return
+
+          // Emit the speaker.updated event since the OS will fallback to the default speaker
+          this.emit('speaker.updated', {
+            previous: {
+              deviceId: disconnectedSpeaker.payload.deviceId,
+              label: disconnectedSpeaker.payload.label,
+            },
+            current: {
+              deviceId: defaultSpeakers.deviceId,
+              label: defaultSpeakers.label,
+            },
+          })
+        }
+      })
     })
   }
 
-  get permissions() {
-    return this.select(({ session }) => {
-      const { authState } = session
-      return (authState as VideoAuthorization)?.room?.scopes ?? []
+  /** @internal */
+  protected override _finalize() {
+    this._screenShareList.clear()
+
+    super._finalize()
+  }
+
+  /** @internal */
+  override async hangup(id?: string) {
+    this._screenShareList.forEach((screenShare) => {
+      screenShare.leave()
     })
+
+    return super.hangup(id)
   }
 
   /**
@@ -166,19 +145,22 @@ export class RoomSessionConnection<
    * exactly when the workers are attached.
    * @internal
    */
-  protected attachPreConnectWorkers() {
+  attachPreConnectWorkers() {
     this.runWorker('memberListUpdated', {
       worker: workers.memberListUpdatedWorker,
     })
   }
 
-  getMemberOverlay(memberId: string) {
-    return this.overlayMap.get(addOverlayPrefix(memberId))
+  /** @internal */
+  getAudioEl() {
+    if (this._audioEl) return this._audioEl
+    this._audioEl = new Audio()
+    this._attachSpeakerTrackListener()
+    return this._audioEl
   }
 
-  /** @deprecated Use {@link startScreenShare} instead. */
-  async createScreenShareObject(opts: CreateScreenShareObjectOptions = {}) {
-    return this.startScreenShare(opts)
+  getMemberOverlay(memberId: string) {
+    return this.overlayMap.get(addOverlayPrefix(memberId))
   }
 
   /**
@@ -259,297 +241,30 @@ export class RoomSessionConnection<
       }
     })
   }
-
-  /**
-   * Allow to add a camera to the room.
-   */
-  addCamera(opts: AddCameraOptions = {}) {
-    const { autoJoin = true, ...video } = opts
-    return this.addDevice({
-      autoJoin,
-      video,
-    })
-  }
-
-  /**
-   * Allow to add a microphone to the room.
-   */
-  addMicrophone(opts: AddMicrophoneOptions = {}) {
-    const { autoJoin = true, ...audio } = opts
-    return this.addDevice({
-      autoJoin,
-      audio,
-    })
-  }
-
-  /**
-   * Allow to add additional devices to the room like cameras or microphones.
-   */
-  async addDevice(opts: AddDeviceOptions = {}) {
-    return new Promise<RoomSessionDevice>(async (resolve, reject) => {
-      const { autoJoin = true, audio = false, video = false } = opts
-      if (!audio && !video) {
-        throw new TypeError(
-          'At least one of `audio` or `video` must be requested.'
-        )
-      }
-
-      const options: BaseConnectionOptions = {
-        ...this.options,
-        localStream: undefined,
-        remoteStream: undefined,
-        audio,
-        video,
-        additionalDevice: true,
-        recoverCall: false,
-        userVariables: {
-          ...(this.options?.userVariables || {}),
-          memberCallId: this.callId,
-          memberId: this.memberId,
-        },
-      }
-
-      const roomDevice = connect<
-        RoomSessionDeviceEvents,
-        RoomSessionDeviceConnection,
-        RoomSessionDevice
-      >({
-        store: this.store,
-        Component: RoomSessionDeviceAPI,
-      })(options)
-
-      roomDevice.once('destroy', () => {
-        roomDevice.emit('room.left')
-        this._deviceList.delete(roomDevice)
-      })
-
-      try {
-        roomDevice.runWorker('childMemberJoinedWorker', {
-          worker: workers.childMemberJoinedWorker,
-          onDone: () => resolve(roomDevice),
-          onFail: reject,
-          initialState: {
-            parentId: this.memberId,
-          },
-        })
-
-        this._deviceList.add(roomDevice)
-        if (autoJoin) {
-          return await roomDevice.join()
-        }
-        return resolve(roomDevice)
-      } catch (error) {
-        this.logger.error('RoomDevice Error', error)
-        reject(error)
-      }
-    })
-  }
-
-  join() {
-    return super.invite<BaseRoomSession<this>>()
-  }
-
-  leave() {
-    return this.hangup()
-  }
-
-  updateSpeaker({ deviceId }: { deviceId: string }) {
-    const prevId = this._audioEl.sinkId as string
-    this.once(
-      // @ts-expect-error
-      `${LOCAL_EVENT_PREFIX}.speaker.updated`,
-      async (newId: string) => {
-        const prevSpeaker = await getSpeakerById(prevId)
-        const newSpeaker = await getSpeakerById(newId)
-
-        const isSame = newSpeaker?.deviceId === prevSpeaker?.deviceId
-        if (!newSpeaker?.deviceId || isSame) return
-
-        // @ts-expect-error
-        this.emit('speaker.updated', {
-          previous: {
-            deviceId: prevSpeaker?.deviceId,
-            label: prevSpeaker?.label,
-          },
-          current: {
-            deviceId: newSpeaker.deviceId,
-            label: newSpeaker.label,
-          },
-        })
-      }
-    )
-
-    return this.triggerCustomSaga<undefined>(audioSetSpeakerAction(deviceId))
-  }
-
-  private _attachSpeakerTrackListener() {
-    if (!supportsMediaOutput()) return
-
-    // @TODO: Stop the watcher when user leave/disconnects
-    createSpeakerDeviceWatcher().then((deviceWatcher) => {
-      deviceWatcher.on('removed', async (data) => {
-        const sinkId = this._audioEl.sinkId
-        const disconnectedSpeaker = data.changes.find((device) => {
-          const payloadDeviceId = device.payload.deviceId
-
-          return (
-            payloadDeviceId === sinkId ||
-            (payloadDeviceId === '' && sinkId === 'default') ||
-            (payloadDeviceId === 'default' && sinkId === '')
-          )
-        })
-        if (disconnectedSpeaker) {
-          // @ts-expect-error
-          this.emit('speaker.disconnected', {
-            deviceId: disconnectedSpeaker.payload.deviceId,
-            label: disconnectedSpeaker.payload.label,
-          })
-
-          /**
-           * In case the currently in-use speaker disconnects, OS by default fallbacks to the default speaker
-           * Set the sink id here to make the SDK consistent with the OS
-           */
-          await this._audioEl.setSinkId?.('')
-
-          const defaultSpeakers = await getSpeakerById('default')
-
-          if (!defaultSpeakers?.deviceId) return
-
-          // Emit the speaker.updated event since the OS will fallback to the default speaker
-          // @ts-expect-error
-          this.emit('speaker.updated', {
-            previous: {
-              deviceId: disconnectedSpeaker.payload.deviceId,
-              label: disconnectedSpeaker.payload.label,
-            },
-            current: {
-              deviceId: defaultSpeakers.deviceId,
-              label: defaultSpeakers.label,
-            },
-          })
-        }
-      })
-    })
-  }
-
-  getAudioEl() {
-    if (this._audioEl) return this._audioEl
-    this._audioEl = new Audio()
-    this._attachSpeakerTrackListener()
-    return this._audioEl
-  }
-
-  /** @internal */
-  override async hangup(id?: string) {
-    this._screenShareList.forEach((screenShare) => {
-      screenShare.leave()
-    })
-    this._deviceList.forEach((device) => {
-      device.leave()
-    })
-
-    return super.hangup(id)
-  }
-
-  /** @internal */
-  protected _finalize() {
-    this._screenShareList.clear()
-    this._deviceList.clear()
-
-    super._finalize()
-  }
-
-  /**
-   * @deprecated Use {@link getLayouts} instead. `getLayoutList` will
-   * be removed in v3.0.0
-   */
-  getLayoutList() {
-    // @ts-expect-error
-    return this.getLayouts()
-  }
-
-  /**
-   * @deprecated Use {@link getMembers} instead. `getMemberList` will
-   * be removed in v3.0.0
-   */
-  getMemberList() {
-    // @ts-expect-error
-    return this.getMembers()
-  }
-
-  /** @internal */
-  protected override getSubscriptions() {
-    const eventNamesWithPrefix = this.eventNames().map((event) => {
-      return `video.${String(event)}`
-    })
-    return validateEventsToSubscribe(
-      eventNamesWithPrefix
-    ) as EventEmitter.EventNames<TEvents & BaseConnectionStateEventTypes>[]
-  }
 }
 
-export const RoomSessionAPI = extendComponent<
-  RoomSessionConnection,
-  RoomMethods
->(RoomSessionConnection, {
-  audioMute: Rooms.audioMuteMember,
-  audioUnmute: Rooms.audioUnmuteMember,
-  videoMute: Rooms.videoMuteMember,
-  videoUnmute: Rooms.videoUnmuteMember,
-  deaf: Rooms.deafMember,
-  undeaf: Rooms.undeafMember,
-  setInputVolume: Rooms.setInputVolumeMember,
-  setOutputVolume: Rooms.setOutputVolumeMember,
-  setMicrophoneVolume: Rooms.setInputVolumeMember,
-  setSpeakerVolume: Rooms.setOutputVolumeMember,
-  setInputSensitivity: Rooms.setInputSensitivityMember,
-  removeMember: Rooms.removeMember,
-  removeAllMembers: Rooms.removeAllMembers,
-  getMembers: Rooms.getMembers,
-  getLayouts: Rooms.getLayouts,
-  setLayout: Rooms.setLayout,
-  setPositions: Rooms.setPositions,
-  setMemberPosition: Rooms.setMemberPosition,
-  hideVideoMuted: Rooms.hideVideoMuted,
-  showVideoMuted: Rooms.showVideoMuted,
-  getRecordings: Rooms.getRecordings,
-  startRecording: Rooms.startRecording,
-  getPlaybacks: Rooms.getPlaybacks,
-  play: Rooms.play,
-  setHideVideoMuted: Rooms.setHideVideoMuted,
-  getMeta: Rooms.getMeta,
-  setMeta: Rooms.setMeta,
-  updateMeta: Rooms.updateMeta,
-  deleteMeta: Rooms.deleteMeta,
-  getMemberMeta: Rooms.getMemberMeta,
-  setMemberMeta: Rooms.setMemberMeta,
-  updateMemberMeta: Rooms.updateMemberMeta,
-  deleteMemberMeta: Rooms.deleteMemberMeta,
-  promote: Rooms.promote,
-  demote: Rooms.demote,
-  getStreams: Rooms.getStreams,
-  startStream: Rooms.startStream,
-  lock: Rooms.lock,
-  unlock: Rooms.unlock,
-  setRaisedHand: Rooms.setRaisedHand,
-  setPrioritizeHandraise: Rooms.setPrioritizeHandraise,
-})
+type BaseRoomSessionEventsHandlerMap = Record<
+  BaseConnectionState,
+  (params: BaseRoomSession<BaseRoomSessionEvents>) => void
+> &
+  Record<MediaEventNames, () => void>
 
-export type RoomSessionObjectEventsHandlerMapping = RoomSessionObjectEvents &
-  BaseConnectionStateEventTypes
+export type BaseRoomSessionEvents = {
+  [k in keyof BaseRoomSessionEventsHandlerMap]: BaseRoomSessionEventsHandlerMap[k]
+}
 
 /** @internal */
-export const createBaseRoomSessionObject = <RoomSessionType>(
-  params: BaseComponentOptions
-): BaseRoomSession<RoomSessionType> => {
+export const createBaseRoomSessionObject = (
+  params: BaseRoomSessionOptions
+): BaseRoomSession<BaseRoomSessionEvents> => {
   const room = connect<
-    RoomSessionObjectEventsHandlerMapping,
-    RoomSessionConnection,
-    BaseRoomSession<RoomSessionType>
+    BaseRoomSessionEvents,
+    BaseRoomSessionConnection<BaseRoomSessionEvents>,
+    BaseRoomSession<BaseRoomSessionEvents>
   >({
     store: params.store,
     customSagas: params.customSagas,
-    Component: RoomSessionAPI,
+    Component: BaseRoomSessionConnection,
   })(params)
 
   return room
