+++ packages/js/src/utils/interfaces/index.ts
@@ -1,2 +1,3 @@
+export * from './base'
 export * from './video'
 export * from './callFabric'
