+++ packages/js/src/utils/interfaces/video.ts
@@ -1,6 +1,5 @@
 import type {
   Rooms,
-  BaseConnectionState,
   VideoLayoutEventNames,
   VideoRoomSessionEventNames,
   VideoRoomEventParams,
@@ -17,7 +16,7 @@ import type {
   VideoPlaybackEventNames,
   RoomSessionRecording,
   RoomSessionPlayback,
-  VideoRoomSessionContract,
+  VideoRoomSessionContract as CoreVideoRoomSessionContract,
   OnlyFunctionProperties,
   MemberListUpdated,
   VideoPositions,
@@ -38,13 +37,14 @@ import type {
   VideoRoomDeviceEventNames,
   VideoLayoutChangedEventParams,
   VideoPosition,
+  BaseConnectionState,
 } from '@signalwire/core'
 import { INTERNAL_MEMBER_UPDATABLE_PROPS } from '@signalwire/core'
-import type { MediaEvent } from '@signalwire/webrtc'
-import type { RoomSession } from '../../RoomSession'
+import type { MediaEventNames } from '@signalwire/webrtc'
 import type { RoomSessionDevice } from '../../RoomSessionDevice'
 import type { RoomSessionScreenShare } from '../../RoomSessionScreenShare'
 import { LocalVideoOverlay, OverlayMap, UserOverlay } from '../../VideoOverlays'
+import { RoomSession } from '../../RoomSession'
 
 /**
  * @privateRemarks
@@ -135,7 +135,7 @@ export type RoomSessionObjectEventsHandlerMap = Record<
     (params: VideoRoomSubscribedEventParams) => void
   > &
   Record<RoomLeft, (params?: RoomLeftEventParams) => void> &
-  Record<MediaEvent, () => void> &
+  Record<MediaEventNames, () => void> &
   Record<
     VideoRoomDeviceUpdatedEventNames,
     (params: DeviceUpdatedEventParams) => void
@@ -154,7 +154,12 @@ export type RoomSessionObjectEventsHandlerMap = Record<
   Record<BaseConnectionState, (params: RoomSession) => void> &
   Record<VideoStreamEventNames, (stream: RoomSessionStream) => void>
 
+// TODO: Remove this and use VideoRoomSessionEvents
 export type RoomSessionObjectEvents = {
+  [k in keyof RoomSessionObjectEventsHandlerMap]: RoomSessionObjectEventsHandlerMap[k]
+}
+
+export type VideoRoomSessionEvents = {
   [k in keyof RoomSessionObjectEventsHandlerMap]: RoomSessionObjectEventsHandlerMap[k]
 }
 
@@ -333,13 +338,14 @@ interface RoomMemberSelfMethodsInterface {
  * flexibility across different objects.
  */
 export interface RoomMethods
-  extends OnlyFunctionProperties<VideoRoomSessionContract> {
+  extends OnlyFunctionProperties<CoreVideoRoomSessionContract> {
   /** @deprecated Use {@link setVideoMuted} instead */
   hideVideoMuted(): Rooms.HideVideoMuted
   /** @deprecated Use {@link setVideoMuted} instead */
   showVideoMuted(): Rooms.ShowVideoMuted
 }
 
+// TODO: Remove this and use VideoRoomSessionContract
 export interface RoomSessionConnectionContract {
   screenShareList: RoomSessionScreenShare[]
   deviceList: RoomSessionDevice[]
@@ -389,6 +395,103 @@ export interface RoomSessionConnectionContract {
    */
   startScreenShare(
     opts?: StartScreenShareOptions
+  ): Promise<RoomSessionScreenShare>
+  /**
+   * Adds a camera device to the room. Using this method, a user can stream
+   * multiple video sources at the same time.
+   *
+   * @param opts - Specify the constraints for the device. In addition, you can
+   * add the `autoJoin` key to specify whether the device should immediately
+   * join the room or joining will be performed manually later. {@link AddCameraOptions}
+   * @returns - {@link RoomSessionDevice}
+   *
+   * @example Adding a specific camera:
+   * ```typescript
+   * await roomSession.addCamera({deviceId: "gOtMHwZdoA6wMlAnhbfTmeRgPAsqa7iw1OwgKYtbTLA="})
+   * ```
+   */
+  addCamera(opts: AddCameraOptions): Promise<RoomSessionDevice>
+  /**
+   * Adds a microphone device to the room. Using this method, a user can stream
+   * multiple video sources at the same time.
+   *
+   * @param opts Specify the constraints for the device. In addition, you can
+   * add the `autoJoin` key to specify whether the device should immediately
+   * join the room or joining will be performed manually later. {@link AddMicrophoneOptions}
+   * @returns - {@link RoomSessionDevice}
+   *
+   * @example Adding a specific microphone:
+   * ```typescript
+   * await roomSession.addMicrophone({deviceId: "PIn/IIDDgBUHzJkhRncv1m85hX1gC67xYIgJvvThB3Q="})
+   * ```
+   */
+  addMicrophone(opts: AddMicrophoneOptions): Promise<RoomSessionDevice>
+  /**
+   * Adds a device to the room. Using this method, a user can stream multiple
+   * sources at the same time. If you need to add a camera device or a
+   * microphone device, you can alternatively use the more specific methods
+   * {@link addCamera} and {@link addMicrophone}.
+   *
+   * @param opts Specify the constraints for the device. In addition, you can
+   * add the `autoJoin` key to specify whether the device should immediately
+   * join the room or joining will be performed manually later. {@link AddDeviceOptions}
+   * @returns - {@link RoomSessionDevice}
+   *
+   * @example Adding any of the microphone devices to the room (duplicate
+   * streams are possible):
+   * ```typescript
+   * await roomSession.addDevice({audio: true})
+   * ```
+   */
+  addDevice(opts: AddDeviceOptions): Promise<RoomSessionDevice>
+  /**
+   * Replaces the current speaker with a different one.
+   *
+   * > 📘
+   * > Some browsers do not support output device selection. You can check by calling {@link WebRTC.supportsMediaOutput}.
+   *
+   * @param opts
+   * @param opts.deviceId id of the new speaker device
+   *
+   * @example Replaces the current speaker:
+   * ```typescript
+   * await room.updateSpeaker({deviceId: "/o4ZeWzroh+8q0Ds/CFfmn9XpqaHzmW3L/5ZBC22CRg="})
+   * ```
+   */
+  updateSpeaker(opts: { deviceId: string }): Promise<undefined>
+}
+
+export interface VideoRoomSessionContract {
+  deviceList: RoomSessionDevice[]
+  interactivityMode: VideoAuthorization['join_as']
+  permissions: VideoAuthorization['scopes']
+  /**
+   * The layout returned from the `layout.changed` event based on the current room layout
+   */
+  currentLayout: VideoLayoutChangedEventParams['layout']
+  /**
+   * The current position of the member returned from the `layout.changed` event
+   */
+  currentPosition: VideoPosition | undefined
+  /**
+   * Joins the room session.
+   */
+  join(options?: BaseRoomSessionJoinParams): Promise<VideoRoomSessionContract>
+  /**
+   * Leaves the room. This detaches all the locally originating streams from the
+   * room.
+   */
+  leave(): Promise<void>
+  /**
+   * Adds a screen sharing instance to the room. You can create multiple screen
+   * sharing instances and add all of them to the room.
+   * @param opts - {@link CreateScreenShareObjectOptions}
+   * @returns - {@link RoomSessionScreenShare}
+   *
+   * @deprecated Use {@link startScreenShare} instead.
+   */
+  createScreenShareObject(
+    opts?: CreateScreenShareObjectOptions
   ): Promise<RoomSessionScreenShare>
   /**
    * Adds a camera device to the room. Using this method, a user can stream
