+++ packages/webrtc/src/utils/interfaces.ts
@@ -1,4 +1,9 @@
 import type { VideoPositions } from '@signalwire/core'
+import {
+  BaseConnectionState,
+  VideoRoomDeviceEventParams,
+  VideoRoomDeviceEventNames,
+} from '@signalwire/core'
 
 export interface ConnectionOptions {
   // TODO: Not used anymore but required for backend
@@ -95,4 +100,30 @@ export interface ConnectionOptions {
 
   layout?: string
   positions?: VideoPositions
+}
+
+export interface OnVertoByeParams {
+  byeCause: string
+  byeCauseCode: string
+  rtcPeerId: string
+  redirectDestination?: string
+}
+
+export type MediaEventNames =
+  | 'media.connected'
+  | 'media.reconnecting'
+  | 'media.disconnected'
+
+type BaseConnectionEventsHandlerMap = Record<
+  BaseConnectionState,
+  (params: any) => void
+> &
+  Record<MediaEventNames, () => void> &
+  Record<
+    VideoRoomDeviceEventNames,
+    (params: VideoRoomDeviceEventParams) => void
+  >
+
+export type BaseConnectionEvents = {
+  [k in keyof BaseConnectionEventsHandlerMap]: BaseConnectionEventsHandlerMap[k]
 }
